re-group widget and workflow tests
