+++ src/ewoksorange/tests/owwidget/test_owwidget_nothread.py
@@ -4,8 +4,8 @@ import time
 import numpy
 from ewokscore.task import Task
 
-from ..gui.owwidgets.meta import ow_build_opts
-from ..gui.owwidgets.threaded import OWEwoksWidgetWithTaskStack
+from ...gui.owwidgets.meta import ow_build_opts
+from ...gui.owwidgets.nothread import OWEwoksWidgetNoThread
 
 
 class Sequential(
@@ -18,18 +18,17 @@ class Sequential(
         self.outputs.thread_id = threading.get_ident()
 
 
-class OWSequential(
-    OWEwoksWidgetWithTaskStack, **ow_build_opts, ewokstaskclass=Sequential
-):
+class OWSequential(OWEwoksWidgetNoThread, **ow_build_opts, ewokstaskclass=Sequential):
     name = "test_OW"
 
 
-def test_owwidget_fifo(qtapp):
-    """Test FIFO execution in a single worker thread."""
+def test_owwidget_nothread(qtapp):
+    """Test sequential execution in the calling thread."""
     widget = OWSequential()
 
     sleep_seconds = 0.5
     values = [0, 1, 2]
+    main_thread = threading.get_ident()
 
     futures = []
     for value in values:
@@ -47,8 +46,8 @@ def test_owwidget_fifo(qtapp):
         thread_ids[value] = outputs["thread_id"].value
 
     times = [results[v] for v in values]
-    not_in_main = [thread_ids[v] != threading.get_ident() for v in values]
+    in_main = [thread_ids[v] == main_thread for v in values]
 
     assert times == sorted(times), "not in order"
     assert (numpy.diff(times) >= sleep_seconds).all(), "not sequential"
-    assert all(not_in_main), "in main thread"
+    assert all(in_main), "not in main thread"
