+++ src/ewoksorange/tests/workflow/test_default_widget.py
@@ -1,7 +1,7 @@
 from ewokscore import Task
 from ewoksutils.import_utils import qualname
 
-from ..gui.workflows.owscheme import ewoks_to_ows
+from ...gui.workflows.owscheme import ewoks_to_ows
 
 
 class Dummy(Task, input_names=["a"], output_names=["b"]):
